move cohortSessions interfaces to slack types
